Split classes from core/events/block_events.js into multiple files
