+++ core/events/events_block_base.js
@@ -1,0 +1,66 @@
+/**
+ * @license
+ * Copyright 2018 Google LLC
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+/**
+ * @fileoverview Base class for all types of block events.
+ * @author fenichel@google.com (Rachel Fenichel)
+ */
+'use strict';
+
+goog.provide('Blockly.Events.BlockBase');
+
+goog.require('Blockly.Events');
+goog.require('Blockly.Events.Abstract');
+goog.require('Blockly.Events.BlockBase');
+goog.require('Blockly.utils.object');
+
+goog.requireType('Blockly.Block');
+
+
+/**
+ * Abstract class for a block event.
+ * @param {!Blockly.Block=} opt_block The block this event corresponds to.
+ *     Undefined for a blank event.
+ * @extends {Blockly.Events.Abstract}
+ * @constructor
+ */
+Blockly.Events.BlockBase = function(opt_block) {
+  Blockly.Events.BlockBase.superClass_.constructor.call(this);
+  this.isBlank = typeof opt_block == 'undefined';
+
+  /**
+   * The block ID for the block this event pertains to
+   * @type {string}
+   */
+  this.blockId = this.isBlank ? '' : opt_block.id;
+
+  /**
+   * The workspace identifier for this event.
+   * @type {string}
+   */
+  this.workspaceId = this.isBlank ? '' : opt_block.workspace.id;
+};
+Blockly.utils.object.inherits(Blockly.Events.BlockBase,
+    Blockly.Events.Abstract);
+
+/**
+ * Encode the event as JSON.
+ * @return {!Object} JSON representation.
+ */
+Blockly.Events.BlockBase.prototype.toJson = function() {
+  var json = Blockly.Events.BlockBase.superClass_.toJson.call(this);
+  json['blockId'] = this.blockId;
+  return json;
+};
+
+/**
+ * Decode the JSON event.
+ * @param {!Object} json JSON representation.
+ */
+Blockly.Events.BlockBase.prototype.fromJson = function(json) {
+  Blockly.Events.BlockBase.superClass_.fromJson.call(this, json);
+  this.blockId = json['blockId'];
+};
